refactor(stream): Move IsTerminal with RawStream
