+++ crates/anstream/src/stream.rs
@@ -1,5 +1,3 @@
-use crate::IsTerminal;
-
 /// Required functionality for underlying [`std::io::Write`] for adaptation
 #[cfg(not(all(windows, feature = "wincon")))]
 pub trait RawStream: std::io::Write + IsTerminal + private::Sealed {}
@@ -22,6 +20,45 @@ impl RawStream for std::io::StderrLock<'static> {}
 impl RawStream for std::fs::File {}
 
 impl RawStream for crate::Buffer {}
+
+pub trait IsTerminal {
+    fn is_terminal(&self) -> bool;
+}
+
+impl IsTerminal for std::io::Stdout {
+    #[inline]
+    fn is_terminal(&self) -> bool {
+        std::io::IsTerminal::is_terminal(self)
+    }
+}
+
+impl IsTerminal for std::io::StdoutLock<'static> {
+    #[inline]
+    fn is_terminal(&self) -> bool {
+        std::io::IsTerminal::is_terminal(self)
+    }
+}
+
+impl IsTerminal for std::io::Stderr {
+    #[inline]
+    fn is_terminal(&self) -> bool {
+        std::io::IsTerminal::is_terminal(self)
+    }
+}
+
+impl IsTerminal for std::io::StderrLock<'static> {
+    #[inline]
+    fn is_terminal(&self) -> bool {
+        std::io::IsTerminal::is_terminal(self)
+    }
+}
+
+impl IsTerminal for std::fs::File {
+    #[inline]
+    fn is_terminal(&self) -> bool {
+        std::io::IsTerminal::is_terminal(self)
+    }
+}
 
 mod private {
     pub trait Sealed {}
